requirements.txt
absl-py==1.4.0
aioice==0.9.0
aiortc==1.5.0
altair==4.2.2
appdirs==1.4.4
asgiref==3.6.0
astunparse==1.6.3
attrs==23.1.0
audioread==3.0.0
av==10.0.0
beautifulsoup4==4.12.2
blinker==1.6.2
bootstrap4==0.1.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
contourpy==1.0.7
cryptography==40.0.2
cssselect==1.2.0
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.7
django-3-jet==1.0.8
django-allauth==0.54.0
django-bootstrap-v5==1.0.11
django-ckeditor==6.5.1
django-crispy-forms==2.0
Django-Gtts==0.4
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-social-share==2.3.0
django-widget-tweaks==1.4.12
dnspython==2.3.0
EasyProcess==1.1
entrypoint2==1.1
entrypoints==0.4
flatbuffers==23.5.9
fonttools==4.40.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-crc32c==1.5.0
google-pasta==0.2.0
googletrans==3.0.0
grpcio==1.54.2
gTTS==2.3.2
h11==0.9.0
h2==3.2.0
h5py==3.8.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ifaddr==0.2.0
importlib-metadata==6.6.0
jax==0.4.10
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
librosa==0.10.0.post2
llvmlite==0.40.1rc1
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
markdown2==2.4.8
MarkupSafe==2.1.2
mathparse==0.1.2
matplotlib==3.7.1
mdurl==0.1.2
ml-dtypes==0.1.0
MouseInfo==0.1.3
msgpack==1.0.5
mss==9.0.1
nltk==3.8.1
numba==0.57.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
Pillow==9.4.0
Pint==0.22
pooch==1.6.0
protobuf==3.20.3
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
PyAutoGUI==0.9.54
pycparser==2.21
pydeck==0.8.1b0
pydub==0.25.1
pyee==10.0.1
PyGetWindow==0.0.9
Pygments==2.15.1
PyJWT==2.6.0
pylibsrtp==0.8.0
pymongo==3.13.0
Pympler==1.0.1
PyMsgBox==1.0.9
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
pyquery==2.0.0
PyRect==0.2.0
pyrsistent==0.19.3
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
python3-openid==3.2.0
pyttsx3==2.90
pytweening==1.0.7
pytz==2023.3
pywin32==306
PyYAML==3.13
readtime==2.0.0
regex==2023.5.5
requests==2.29.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.3.5
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
SpeechRecognition==3.10.0
SQLAlchemy==1.2.19
sqlparse==0.4.3
streamlit==1.22.0
streamlit-option-menu==0.3.4
streamlit-webrtc==0.45.1
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==5.0.1
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.3.4
wikipedia==1.4.0
Wikipedia-API==0.5.8
wrapt==1.14.1
zipp==3.15.0
